+++ stimulus/controller_tb.v
@@ -31,10 +31,17 @@ module controller_tb;
 
 parameter SYSCLK_PERIOD = 1;// 1000MHZ
 
+/* --- external signals --- */
 reg SYSCLK;
 reg NSYSRESET;
-wire [7:0] IO_Target;
+reg [1:0] type;
+
+wire [7:0] IO_Target, IO_Host;
 wire nCE, CLE, ALE, nWE, nRE, nWP;
+
+/* --- internal signals --- */
+reg [7:0] buff;
+reg [7:0] count;
 
 initial
 begin
@@ -54,7 +61,31 @@ begin
     NSYSRESET = 1'b1;
 end
 
+//////////////////////////////////////////////////////////////////////
+// Host Simulator
+//////////////////////////////////////////////////////////////////////
+initial begin
+    #(SYSCLK_PERIOD * 10 )
+    
+    type = 2'b00;
+    buff = 8'hFF;
+    count = 8'b0;
+    
+end
 
+always @(posedge SYSCLK) begin
+    if (buff != 1'bz) begin
+        count = count +1;
+    end
+    if (count % 20 == 0) begin
+        buff = buff+1;
+    end
+    if (count % 50 == 0) begin
+        type = type+1;
+    end
+end
+
+assign IO_Host = buff;
 //////////////////////////////////////////////////////////////////////
 // Clock Driver
 //////////////////////////////////////////////////////////////////////
@@ -77,9 +108,10 @@ MT29F8G08ABACAWP MT29F8G08ABACAWP_0 (
     .nWE(nWE),
     .nRE(nRE),
     .nWP(nWP),
+    .TYPE(type),
 
     // Inouts
-    .IOH( ),
+    .IOH( IO_Host ),
     .IOT( IO_Target )
 
 );