+++ changelog.md
@@ -4,6 +4,8 @@ The NAND flash controller for Micron MT29F8G08ABACAWP by verilog.
 -----------------------------------
 ## Unreleased v0.0.3 (2023.06.##)
 ### New
+1. create an always block to deal with type and IOH from host
+2. create host simulator in testbench
 ### Changed
 ### Removed
 ### Deprecated
@@ -13,11 +15,11 @@ The NAND flash controller for Micron MT29F8G08ABACAWP by verilog.
 ### Reference
 
 -----------------------------------
-## Unreleased v0.0.2 (2023.06.13)
+## NFC v0.0.2 (2023.06.13)
 ### New
 1. create task `address_cycle`
 2. create task `data_cycle`
-3. create input `MODE`
+3. create input `TYPE`
 4. 
 -----------------------------------
 ## NFC v0.0.1 (2023.06.13)